FIX: fixes to database creation
