+++ pyfibre/io/utilities.py
@@ -172,6 +172,35 @@ def python_to_numpy_recursive(dictionary):
     return dictionary
 
 
+def check_string(string, pos, sep, word):
+    """Checks index 'pos' of 'string' seperated by 'sep' for substring 'word'
+    If present, removes 'word' and returns amended string
+    """
+
+    if sep in string:
+        temp_string = string.split(sep)
+        if temp_string[pos] == word:
+            temp_string.pop(pos)
+        string = sep.join(temp_string)
+
+    return string
+
+
+def check_file_name(file_name, file_type="", extension=""):
+    """
+    check_file_name(file_name, file_type="", extension="")
+
+    Checks file_name for file_type or extension. If present, returns
+    amended file_name without extension or file_type
+
+    """
+
+    file_name = check_string(file_name, -1, '.', extension)
+    file_name = check_string(file_name, -1, '_', file_type)
+
+    return file_name
+
+
 def replace_ext(file_name, extension):
     """If an extension exists on file_name,
     replace it with new extension. Otherwise
@@ -183,6 +212,22 @@ def replace_ext(file_name, extension):
         file_name = path + f'.{extension}'
 
     return file_name
+
+
+def get_file_names(prefix):
+
+    image_name = os.path.basename(prefix)
+    working_dir = (
+        f"{os.path.dirname(prefix)}/{image_name}"
+        "-pyfibre-analysis")
+
+    data_dir = working_dir + '/data/'
+    fig_dir = working_dir + '/fig/'
+
+    filename = data_dir + image_name
+    figname = fig_dir + image_name
+
+    return working_dir, data_dir, fig_dir, filename, figname
 
 
 def save_json(data, file_name):
