+++ requirements.txt
@@ -8,4 +8,5 @@ pandas==1.0.0
 opencv-python==3.4.5.20
 imageio==2.4.1
 xlwt==1.3.0
+xlrd==1.2.0
 tables==3.5.2
